# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 股票/基金行情平台 API
  description: 股票与基金行情平台后端服务接口文档，提供基金搜索、股票搜索和市场行情等功能；旧预测路径仅作为已废弃兼容端点保留。错误码体系参见 ErrorCode schema。
  version: 1.0.0
servers:
  - url: http://localhost:5070
    description: 本地开发服务器
...
security: []
paths:
...
  /api/v1/search:
    get:
      tags: [搜索]
      summary: 统一搜索
      description: 同时搜索基金和股票，返回匹配结果
      operationId: unifiedSearch
      parameters:
        - name: q
          in: query
          required: true
          description: 搜索关键词
          schema:
            type: string
            maxLength: 100
        - name: types
          in: query
          required: false
          description: 搜索类型，逗号分隔，可选值为 fund、stock，为空时搜索全部
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: 页码，默认1
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          required: false
          description: 每页数量，默认10，最大50
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        "200":
          description: 搜索结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UnifiedSearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/funds/search:
    get:
      tags: [基金]
      summary: 基金搜索
      description: 根据条件搜索基金列表，支持关键词、类型、公司等多维度筛选
      operationId: searchFunds
      parameters:
        - name: keyword
          in: query
          required: false
          description: 搜索关键词（基金代码或名称）
          schema:
            type: string
        - name: type
          in: query
          required: false
          description: 基金类型筛选
          schema:
            type: string
        - name: company
          in: query
          required: false
          description: 基金公司筛选
          schema:
            type: string
        - name: risk_level
          in: query
          required: false
          description: 风险等级筛选
          schema:
            type: string
        - name: manager
          in: query
          required: false
          description: 基金经理筛选
          schema:
            type: string
        - name: return_min
          in: query
          required: false
          description: 收益率下限
          schema:
            type: number
            format: double
        - name: return_max
          in: query
          required: false
          description: 收益率上限
          schema:
            type: number
            format: double
        - name: sort_by
          in: query
          required: false
          description: 排序字段
          schema:
            type: string
        - name: sort_order
          in: query
          required: false
          description: 排序方向（asc/desc）
          schema:
            type: string
            enum: [asc, desc]
        - name: page
          in: query
          required: false
          description: 页码
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          required: false
          description: 每页数量
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        "200":
          description: 基金搜索结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/FundSearchData"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/funds/sync:
    post:
      tags: [基金]
      summary: 同步基金数据
      description: 从配置的数据源同步基金数据，需要管理员令牌认证
      operationId: syncFunds
      security:
        - adminToken: []
          csrfCookie: []
          csrfHeader: []
      responses:
        "200":
          description: 同步结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/FundSyncResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/market/ranking/{type}:
    get:
      tags: [市场]
      summary: 基金排行榜
      description: 获取基金涨幅或跌幅排行榜
      operationId: marketRanking
      parameters:
        - name: type
          in: path
          required: true
          description: 排行类型，gainers为涨幅榜，losers为跌幅榜
          schema:
            type: string
            enum: [gainers, losers]
        - name: size
          in: query
          required: false
          description: 返回数量，默认10
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        "200":
          description: 基金排行列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/FundRankingItem"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/market/index/{code}/kline:
    get:
      tags: [市场]
      summary: 指数 K 线
      description: 获取指定市场指数的历史 K 线数据
      operationId: indexKline
      parameters:
        - name: code
          in: path
          required: true
          description: 指数代码
          schema:
            type: string
        - name: count
          in: query
          required: false
          description: 返回条数，默认 120，最大 500
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 120
      responses:
        "200":
          description: 指数 K 线数据
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/IndexKlinePoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/market/index/{code}/minute:
    get:
      tags: [市场]
      summary: 指数分时
      description: 获取指定市场指数的当日分时数据
      operationId: indexMinute
      parameters:
        - name: code
          in: path
          required: true
          description: 指数代码
          schema:
            type: string
      responses:
        "200":
          description: 指数分时数据
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/IndexMinutePoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /api/v1/predict/{fundCode}:
    get:
      tags: [预测]
      summary: 基金预测兼容端点（已废弃）
      description: 预测服务已迁移至独立项目。此路径仅为兼容旧调用保留，合法基金代码固定返回 410 Gone。
      operationId: predictFund
      deprecated: true
      parameters:
        - name: fundCode
          in: path
          required: true
          description: 基金代码，6位数字
          schema:
            type: string
            pattern: "^[0-9]{6}$"
      responses:
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          description: 预测服务已迁移，当前端点已废弃
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
              example:
                code: -1
                message: "预测服务已迁移，此接口已废弃"
                data: null
  /api/v1/watchlist/quotes:
    post:
      tags: [自选股]
      summary: 自选基金行情
      description: 批量获取自选基金的实时行情数据，最多支持50个基金代码
      operationId: watchlistQuotes
      security:
        - csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistQuoteRequest"
      responses:
        "200":
          description: 自选基金行情列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/WatchlistItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/fund/{fundCode}/detail:
    get:
      tags: [基金]
      summary: 基金详情
      description: 获取基金详细信息，包括基本信息、行情、业绩表现、基金经理、持仓、风险指标等
      operationId: fundDetail
      parameters:
        - name: fundCode
          in: path
          required: true
          description: 基金代码，6位数字
          schema:
            type: string
            pattern: "^[0-9]{6}$"
      responses:
        "200":
          description: 基金详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/FundDetailData"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/stock/{stockCode}/predict:
    get:
      tags: [预测]
      summary: 股票预测兼容端点（已废弃）
      description: 预测服务已迁移至独立项目。此路径仅为兼容旧调用保留，合法股票代码固定返回 410 Gone。
      operationId: predictStock
      deprecated: true
      parameters:
        - name: stockCode
          in: path
          required: true
          description: 股票代码，6位数字
          schema:
            type: string
            pattern: "^[0-9]{6}$"
      responses:
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          description: 预测服务已迁移，当前端点已废弃
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
              example:
                code: -1
                message: "预测服务已迁移，此接口已废弃"
                data: null
...
components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
      description: 管理员令牌，通过 Authorization 请求头传递 Bearer Token
    csrfCookie:
      type: apiKey
      in: cookie
      name: csrf_token
      description: 先通过任意 GET 请求获取 HttpOnly CSRF Cookie
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 使用 GET 响应头返回的令牌，值必须与 csrf_token Cookie 一致
  schemas:
    APIResponse:
      type: object
      description: 通用API响应包装
      properties:
        code:
          type: integer
          description: 状态码，0表示成功，正数为业务错误码，-1为通用错误
          example: 0
        message:
          type: string
          description: 状态消息
          example: success
        data:
          description: 响应数据
...
    FundItem:
      type: object
      description: 基金基本信息
      properties:
        fund_code:
          type: string
          description: 基金代码
          example: "000001"
        fund_name:
          type: string
          description: 基金名称
          example: 华夏成长混合
        fund_type:
          type: string
          description: 基金类型
          example: 混合型
        pinyin_abbr:
          type: string
          description: 拼音缩写
        pinyin_full:
          type: string
          description: 拼音全拼
        company:
          type: string
          description: 基金公司
        manager:
          type: string
          description: 基金经理
        latest_nav:
          type: number
          format: double
          description: 最新净值
        cumulative_nav:
          type: number
          format: double
          description: 累计净值
        return_1m:
          type: number
          format: double
          description: 近1月收益率
        return_3m:
          type: number
          format: double
          description: 近3月收益率
        return_6m:
          type: number
          format: double
          description: 近6月收益率
        return_1y:
          type: number
          format: double
          description: 近1年收益率
        return_3y:
          type: number
          format: double
          description: 近3年收益率
        risk_level:
          type: string
          description: 风险等级
        inception_date:
          type: string
          description: 成立日期
        quote_date:
          type: string
          description: 行情日期
        quote_source:
          type: string
          description: 行情来源
    FundSearchData:
      type: object
      description: 基金搜索结果
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FundItem"
        total:
          type: integer
          description: 总数
        page:
          type: integer
          description: 当前页码
        size:
          type: integer
          description: 每页数量
...
    FundSyncResult:
      type: object
      description: 基金同步结果
      properties:
        source:
          type: string
          description: 数据来源
        stored_path:
          type: string
          description: 存储路径
        imported:
          type: integer
          description: 导入数量
        total:
          type: integer
          description: 总数量
        updated_at:
          type: string
          description: 更新时间
    FundRankingItem:
      type: object
      description: 基金排行项
      properties:
        rank:
          type: integer
          description: 排名
        fund_code:
          type: string
          description: 基金代码
        fund_name:
          type: string
          description: 基金名称
        fund_type:
          type: string
          description: 基金类型
        change_pct:
          type: number
          format: double
          description: 涨跌幅百分比
        estimated_nav:
          type: number
          format: double
          description: 估算净值
        quote_date:
          type: string
          description: 行情日期
        quote_source:
          type: string
          description: 行情来源
    WatchlistQuoteRequest:
      type: object
      description: 自选基金行情请求
      required:
        - codes
      properties:
        codes:
          type: array
          items:
            type: string
            pattern: "^[0-9]{6}$"
          minItems: 1
          maxItems: 50
          description: 基金代码列表，每个代码为6位数字，最多50个
    WatchlistItem:
      type: object
      description: 自选基金行情项
      properties:
        fund_code:
          type: string
          description: 基金代码
        fund_name:
          type: string
          description: 基金名称
        fund_type:
          type: string
          description: 基金类型
        estimated_nav:
          type: number
          format: double
          description: 估算净值
        change_pct:
          type: number
          format: double
          description: 涨跌幅百分比
        direction:
          type: string
          description: 涨跌方向
          enum: [up, down, flat]
        added_at:
          type: integer
          format: int64
          description: 添加时间戳
        quote_date:
          type: string
          description: 行情日期
        quote_source:
          type: string
          description: 行情来源
...
    IndexKlinePoint:
      type: object
      description: 指数 K 线数据点
      properties:
        date:
          type: string
          description: 日期
        open:
          type: number
          format: double
          description: 开盘点位
        close:
          type: number
          format: double
          description: 收盘点位
        high:
          type: number
          format: double
          description: 最高点位
        low:
          type: number
          format: double
          description: 最低点位
        volume:
          type: integer
          format: int64
          description: 成交量
        amount:
          type: number
          format: double
          description: 成交额
    IndexMinutePoint:
      type: object
      description: 指数或股票分时数据点
      properties:
        time:
          type: string
          description: 分时时间
        price:
          type: number
          format: double
          description: 最新价格或点位
        avg_price:
          type: number
          format: double
          description: 均价
        volume:
          type: integer
          format: int64
          description: 成交量
...
    FundDetailData:
      type: object
      description: 基金详情数据
      properties:
        basic:
          $ref: "#/components/schemas/FundItem"
        quote:
          $ref: "#/components/schemas/FundItem"
        performance:
          $ref: "#/components/schemas/FundPerformanceData"
        manager:
          $ref: "#/components/schemas/FundManagerInfo"
        portfolio:
          $ref: "#/components/schemas/FundPortfolioData"
        risk:
          $ref: "#/components/schemas/FundRiskMetrics"
    FundPerformanceData:
      type: object
      description: 基金业绩数据
      properties:
        nav_history:
          type: array
          items:
            $ref: "#/components/schemas/NAVPoint"
          description: 净值历史
        return_1m:
          type: number
          format: double
          description: 近1月收益率
        return_3m:
          type: number
          format: double
          description: 近3月收益率
        return_6m:
          type: number
          format: double
          description: 近6月收益率
        return_1y:
          type: number
          format: double
          description: 近1年收益率
        return_3y:
          type: number
          format: double
          description: 近3年收益率
    NAVPoint:
      type: object
      description: 净值数据点
      properties:
        date:
          type: string
          description: 日期
        nav:
          type: number
          format: double
          description: 单位净值
        cumulative_nav:
          type: number
          format: double
          description: 累计净值
        change_pct:
          type: number
          format: double
          description: 涨跌幅百分比
    FundManagerInfo:
      type: object
      description: 基金经理信息
      properties:
        name:
          type: string
          description: 基金经理姓名
        tenure_days:
          type: integer
          description: 任职天数
        managed_size:
          type: string
          description: 管理规模
        fund_count:
          type: integer
          description: 管理基金数量
        bio:
          type: string
          description: 简介
    FundPortfolioData:
      type: object
      description: 基金持仓数据
      properties:
        top_holdings:
          type: array
          items:
            $ref: "#/components/schemas/HoldingItem"
          description: 前十大持仓
        sector_allocation:
          type: array
          items:
            $ref: "#/components/schemas/SectorItem"
          description: 行业配置
    HoldingItem:
      type: object
      description: 持仓项
      properties:
        name:
          type: string
          description: 持仓名称
        code:
          type: string
          description: 持仓代码
        ratio:
          type: number
          format: double
          description: 持仓比例
...
    FundRiskMetrics:
      type: object
      description: 基金风险指标
      properties:
        volatility_1y:
          type: number
          format: double
          description: 1年波动率
        max_drawdown_1y:
          type: number
          format: double
          description: 1年最大回撤
        sharpe_1y:
          type: number
          format: double
          description: 1年夏普比率
        beta_1y:
          type: number
          format: double
          description: 1年Beta系数
    StockItem:
      type: object
      description: 股票基本信息
      properties:
        stock_code:
          type: string
          description: 股票代码
          example: "600519"
        stock_name:
          type: string
          description: 股票名称
          example: 贵州茅台
        market:
          type: string
          description: 所属市场
          example: SH
        industry:
          type: string
          description: 所属行业
        list_date:
          type: string
          description: 上市日期
        total_shares:
          type: number
          format: double
          description: 总股本
        float_shares:
          type: number
          format: double
          description: 流通股本
        current_price:
          type: number
          format: double
          description: 当前价格
        change_pct:
          type: number
          format: double
          description: 涨跌幅百分比
        volume:
          type: number
          format: double
          description: 成交量
        amount:
          type: number
          format: double
          description: 成交额
        turnover_rate:
          type: number
          format: double
          description: 换手率
        pe_ratio:
          type: number
          format: double
          description: 市盈率
        pb_ratio:
          type: number
          format: double
          description: 市净率
        total_mv:
          type: number
          format: double
          description: 总市值
        pinyin:
          type: string
          description: 拼音
        pinyin_alt:
          type: string
          description: 拼音备选
    StockSearchData:
      type: object
      description: 股票搜索结果
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StockItem"
        total:
          type: integer
          description: 总数
        page:
          type: integer
          description: 当前页码
        size:
          type: integer
          description: 每页数量
...
    UnifiedSearchResponse:
      type: object
      description: 统一搜索响应
      properties:
        query:
          type: string
          description: 搜索关键词
        funds:
          $ref: "#/components/schemas/FundSearchData"
        stocks:
          $ref: "#/components/schemas/StockSearchData"
        suggestions:
          type: array
          items:
            type: string
          description: 搜索建议
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            code: -1
            message: "查询参数格式错误"
            data: null
    Unauthorized:
      description: 未授权，管理员令牌无效或缺失
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            code: -1
            message: "未提供管理员令牌"
            data: null
    Forbidden:
      description: CSRF Cookie 或请求头缺失、不一致
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            code: -1
            message: "CSRF token 验证失败"
            data: null
...
    InternalServerError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            code: -1
            message: "服务器内部错误"
            data: null
    ServiceUnavailable:
      description: 服务暂时不可用
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            code: -1
            message: "服务暂时不可用"
            data: null